:white_check_mark: Write tests for release note dialog
