+++ src/organisms/TopBar/Resources/ReleaseNotesDialog/ReleaseNotesDialog.styles.ts
